workshops/2025-05/classification-final/pyproject.toml
[project]
name = "large-scale-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "baml-py==0.82.0",
    "numpy>=2.2.4",
    "openai>=1.70.0",
    "python-dotenv>=1.1.0",
]
